+++ hutool-cron/src/main/java/cn/hutool/cron/pattern/parser/PartParser.java
@@ -3,6 +3,7 @@ package cn.hutool.cron.pattern.parser;
 import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.date.Month;
 import cn.hutool.core.date.Week;
+import cn.hutool.core.lang.Console;
 import cn.hutool.core.util.NumberUtil;
 import cn.hutool.core.util.StrUtil;
 import cn.hutool.cron.CronException;
@@ -74,7 +75,7 @@ public class PartParser {
 		}
 
 		final List<Integer> values = parseArray(value);
-		if (values.size() == 0) {
+		if (values.isEmpty()) {
 			throw new CronException("Invalid part value: [{}]", value);
 		}
 
@@ -251,6 +252,12 @@ public class PartParser {
 		// 周日可以用0或7表示，统一转换为0
 		if(Part.DAY_OF_WEEK.equals(this.part) && Week.SUNDAY.getIso8601Value() == i){
 			i = Week.SUNDAY.ordinal();
+		}
+
+		// issue#I7SMP7
+		// 年的形式中，如果类似于*/2，不做范围检查
+		if(Part.YEAR.equals(this.part)){
+			return i;
 		}
 
 		return part.checkValue(i);